src/neural-roots-backend/requirements.txt
django
psycopg2-binary==2.9.6
gunicorn==21.2.0
djangorestframework
